Update lesson.md
added of to the sentence.
A little explanation of how to import more than one component defined in one file would be helpful here. In this example we only have one component and one file and use the name of the file as a component name but if we had more than one component  per file, as the text suggests, how would that look.