Adding Code of Conduct to Solid-Site
While there is a Code of Conduct provided in the Solid repository, the Solid repository is just another repository of the SolidJS collection. With such, it is entirely likely that a new contributor, such as myself, goes to a repository that sounds interesting (such as Solid-Site), contributes, and never once in their experience in contributing to the SolidJS repositories do they encounter critical documentation (such as the Code of Conduct). 

I believe that the proposed addition of this file signifies an important step in the direction towards ensuring contributors are kept informed of critical information relating to correctly contributing-to and abiding the expectations of contributions to SolidJS repositories.