Increase bottom padding on doc's sidebar
A browser default link preview/tooltip shows on the bottom left of the viewport whenever the user has their cursor over a link.

Issue:
When user engaged in scrolling on left sidebar/table of contents, inevitably users will hover on links during scroll and raise the browser tooltip.
Upon scrolling to the end, without bottom padding of the sidebar, browser's tooltip actually blocks the last item to be viewed. 
This impedes the fluidity of quickly glancing over/scrolling over all the items.

Solution:
This tiny change adds the bottom padding which gives enough spacing so the tooltip no longer obstruct the listed items when user scrolls to the end. Additionally it adds a sense of balance with appropriate spacing with the edge of the window.